+++ src/__init__.py
@@ -1,3 +1,5 @@
 from . utils import get_package_version
 
 __version__ = get_package_version()
+
+PROGRAM_NAME = 'MosaicTool'
